Added support for package-level queries (and a README)
I noticed that package-level queries were not supported, with a vague note about
how Gerrit supports arbitrarily deep repos making this impossible.  It's actually
pretty simple to support this: you just find the package with the greatest depth
that matches the request.

If you have a Gerrit setup with packages that have overlapping names (for example,
you have both "a" and "a/b" projects at the same time), then when a request for
package "a/b" is made, repo "a/b" will be preferred over "a" (even if "a" actually
had a package "b").  Just because Gerrit allows you to do crazy things with your
names doesn't mean that you should actually take it up on that offer.

Anyway, we're using this now in production and it has made it so that `go dep` works
with our private Gerrit instance.

Change-Id: I5c5daa4984a5993aec39da60ea8cca2d3ab94415