Use anonymous project clone URLs
The previous URL prefixed the project name with "/a/", in effect
requiring the user to authenticate with Gerrit before issuing `go get`
and related commands.

Enforcement of authentication is prohibitive for those wanting to simply
clone a project for the first time without the prospect of contribution,
and additionally problematic when a Gerrit hosted repo is being
installed as another project's dependency by a package manager.

Change-Id: I8b5bfbf3019dc7d1920bbfab048412f29962d36e