Update patch set 5
Patch Set 5: Code-Review+1

Thanks for the review, Marco!

I've refactored the patch to work conditionally based on whether the anonymous user has read access to "ref/heads/" and included some tests. This was my first experience with just about every part of the toolchain, so I hope I did most of it correctly. :)

Note that the logic of testing for read access to "ref/heads/" may not be 100% as it's possible for a repo to allow read on all but the default branch, but that seemed like a very rare edge case not worth solving for.

Patch-set: 5
Reviewer: Gerrit User 1069829 <1069829@173816e5-2b9a-37c3-8a2e-48639d4f1153>
Label: Code-Review=+1