Update patch set 3
Patch Set 3:

This patch makes the repo lookup succeed, i.e.

 curl https://gerrit.corp.example.com/a/my/project

will now return 200, but the meta tag will look like this:

 <meta name="go-import" content="gerrit.corp.example.com/my/project git https://gerrit.corp.example.com/a/my/project"/>

This won't be acceptable to 

 go get gerrit.corp.example.com/a/my/project

since that import path given to "go get" doesn't match the import path in the meta tag, and

 go get gerrit.corp.example.com/my/project

won't work either because it gets caught in the login redirection. Am I misunderstanding something, or do we need to propagate the /a prefix to the returned import path if it was present in the path of the go-get=1 request?

Patch-set: 3
CC: Gerrit User 1021683 <1021683@173816e5-2b9a-37c3-8a2e-48639d4f1153>