Re-add support for projects with anonymous HTTP access
This was originally added by go-import#190630, but was reverted by the
merging of go-import#241958 into master.

Enforcing authentication in all cases is problematic for those who wish
to make their public Gerrit hosted Go projects easily installable via
the standard Go toolchain, either as standalone applications or as
libraries.

Anonymous repo URLs are now used as long as the anonymous user is
granted read access to the project's `refs/heads/*`.

Change-Id: I15aa7251483877aae6552df0d11032d63af4b9e2