Fix malfunctioning tests
Fix testDoFilterWithExistingProjectAndPackage test by adding a missing
line which adds a mock project to the cache. Without this line, the test
will assume that the entire string used is a valid project which is not
the case (The extra is meant to represent a package, not a project).
This fix is based on the changes made on the parent commit.

Fix testDoFilterWithNonExistingProject test by using ofNullable option.
Optional.of does not work with null values and will throw a
NullPointerException when the argument is null. ofNullable is designed
for arguments that may or may not be null and therefore allows the usage
of null in this specific test.

Bug: Issue 12454
Change-Id: I356751bec4ccb6f44226219c9ae7b41e7dc02cd7